requirements.txt
apscheduler
Babel
Flask
Flask-APScheduler
Flask-Babel
Flask-Bootstrap
Flask-HTTPAuth
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
Flask-WTF
Jinja2
Mako
MarkupSafe
PyJWT
SQLAlchemy
WTForms
Werkzeug
alembic
blinker
certifi
chardet
click
dominate
elasticsearch
flask-admin
flask-simpleldap
guess_language-spirit
gunicorn
idna
itsdangerous
mysql
mysql-connector-python
openpyxl
pandas
pymysql
python-dateutil
python-dotenv
python-editor
python-ldap
python-redmine
pytz
redis
requests
rq
six
urllib3
visitor
xlrd
